+++ nengo_dl/tests/test_neurons.py
@@ -53,13 +53,49 @@ def test_soft_lif(Simulator, sigma, seed):
 
 
 @pytest.mark.parametrize(
-    "neuron_type", (nengo.LIFRate, SoftLIFRate))
+    "neuron_type", (nengo.LIFRate, nengo.RectifiedLinear, SoftLIFRate))
 def test_neuron_gradients(Simulator, neuron_type, seed):
     with nengo.Network(seed=seed) as net:
         a = nengo.Node(output=[0])
         b = nengo.Ensemble(50, 1, neuron_type=neuron_type())
+        c = nengo.Ensemble(50, 1, neuron_type=neuron_type(amplitude=0.1))
         nengo.Connection(a, b, synapse=None)
-        nengo.Probe(b)
+        nengo.Connection(b, c, synapse=None)
+        nengo.Probe(c)
 
     with Simulator(net, seed=seed) as sim:
         sim.check_gradients()
+
+
+@pytest.mark.parametrize("rate, spiking", [
+    (nengo.RectifiedLinear, nengo.SpikingRectifiedLinear),
+    (nengo.LIFRate, nengo.LIF)])
+def test_spiking_swap(Simulator, rate, spiking, seed):
+    grads = []
+    for neuron_type in [rate, spiking]:
+        with nengo.Network(seed=seed) as net:
+            a = nengo.Node(output=[0])
+            b = nengo.Ensemble(50, 1, neuron_type=neuron_type())
+            c = nengo.Ensemble(50, 1, neuron_type=neuron_type(amplitude=0.1))
+            nengo.Connection(a, b, synapse=None)
+            nengo.Connection(b, c, synapse=None)
+            p = nengo.Probe(c)
+            p_neuron = nengo.Probe(c.neurons)
+
+        with Simulator(net) as sim:
+            grads.append(sim.sess.run(
+                tf.gradients(sim.tensor_graph.probe_arrays[p],
+                             tf.trainable_variables()),
+                feed_dict=sim._fill_feed(10, {}, training=True)))
+
+            sim.soft_reset()
+            sim.run(0.5)
+
+        # check that the normal output is unaffected by the swap logic
+        with nengo.Simulator(net) as sim2:
+            sim2.run(0.5)
+
+            assert np.allclose(sim.data[p_neuron], sim2.data[p_neuron])
+
+    # check that the gradients match
+    assert all(np.allclose(g0, g1) for g0, g1 in zip(*grads))